pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "swfl-arrest-scrapers"
version = "2.0.0"
description = "Florida county jail arrest scraping system"
requires-python = ">=3.10"

dependencies = [
    "DrissionPage>=4.0",
    "google-auth>=2.0",
    "google-auth-oauthlib>=1.0",
    "google-api-python-client>=2.0",
    "gspread>=5.0",
    "python-dotenv>=1.0",
    "pyyaml>=6.0",
    "requests>=2.28",
    "beautifulsoup4>=4.12",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
addopts = "-v --tb=short"

[tool.setuptools.packages.find]
include = ["core*", "counties*", "scripts*"]
